requirements.txt
argcomplete==3.2.3
asgiref==3.7.2
certifi==2024.2.2
channels==4.1.0
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.15.1
fonttools==4.47.2
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pillow==10.2.0
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
sdkmanager==0.6.7
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==1.26.18
websockets==12.0